Flag to allow certain species to always have perfect IVs (like Partner Pikachu and Eevee)

+++ include/pokemon.h
@@ -323,7 +323,7 @@ struct BaseStats
             u8 safariZoneFleeRate;
             u8 bodyColor : 7;
             u8 noFlip : 1;
-            u8 flags;
+            u16 flags;
 };
 
 #include "constants/battle_config.h"
